s3: add support for dreamhost and bucket names in the domain
S3 supports two ways to address bucket names: via a common
domain name plus a path, and via a domain named after the
user bucket. This adds support to the latter convention.

It also includes support for talking to DreamHost's S3-like
endpoint, and fixes a bug in the tests.