magic checksum changes
I'm just changing the checksum to what it expects during my run.

This type of test requires many intimate details of the client
implementaiton, which I don't have. I assume it's the test that's broken
and not the code because the code doesn't seem to have changed, and it's
working anyway. sorry!