Fix complete upload bug in s3 multipart upload.
The multipart upload operation of s3 works by a sequence of calls.
The last one is multipart.Complete. The s3 semantics are rather
kludgy: if they can successfully start the process of completing a
multipart upload, they return HTTP 200, and periodically write a space
into the response body to maintain the connection. If the complete operation
fails during processing, that's passed to the client by putting an
Error element into the response body.

The go implementation of the multipart.Complete method wasn't checking
the message body, and was just declaring success when it saw the 200.